[#2060] Move serialization lookups into a separate file
